pyproject.toml
[tool.poetry]
name = "text-analysis"
version = "0.1.0"
description = "RAG pipeline application with streaming inference API"
authors = []
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = { version = "^0.34", extras = ["standard"] }
langchain-core = "^0.3"
langchain-huggingface = "^0.2"
pydantic-settings = "^2.7"
sqlalchemy = { version = "^2.0", extras = ["asyncio"] }
asyncpg = "^0.30"
httpx = "^0.28"
beautifulsoup4 = "^4.12"
lxml = "^5.3"
apscheduler = "^3.11"
langdetect = "^1.0"
langchain-text-splitters = "^0.3"
pgvector = "^0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
